testing startpage and spotifyLogin
