+++ .eslintrc.js
@@ -1,1 +1,1 @@
-module.exports = { "extends": "airbnb-base" };
+module.exports = { 'extends': 'airbnb-base' };
