+++ package.json
@@ -11,5 +11,10 @@
     "express": "^4.15.4",
     "querystring": "^0.2.0",
     "request": "^2.81.0"
+  },
+  "devDependencies": {
+    "eslint": "^5.5.0",
+    "eslint-config-airbnb-base": "^13.1.0",
+    "eslint-plugin-import": "^2.14.0"
   }
 }
