+++ server.js
@@ -5,9 +5,9 @@ require('dotenv').config()
 
 let app = express()
 
-let redirect_uri = 
-  process.env.REDIRECT_URI || 
-  'http://localhost:8888/callback'
+let redirect_uri =
+  process.env.REDIRECT_URI ||
+  'http://localhost:8888/callback';
 
 app.get('/login', function(req, res) {
   res.redirect('https://accounts.spotify.com/authorize?' +
